% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_bets.R
\name{effective_bets}
\alias{effective_bets}
\title{Effective Number of Bets}
\usage{
effective_bets(b, sigma, t)
}
\arguments{
\item{b}{A vector of exposures.}

\item{sigma}{A \code{n x n} covariance matrix.}

\item{t}{A \code{n x n} torsion matrix.}
}
\value{
A \code{list} with the effective number of bets - enb - and the diversification probability distribution - p.
}
\description{
Computes the diversification probability distribution and the effective number
of bets an allocation.
}
\details{
Transcripted from the MATLAB code in \url{https://la.mathworks.com/matlabcentral/fileexchange/43245-portfolio-diversi-cation-based-on-optimized-uncorrelated-factors?s_tid=prof_contriblnk}.
}
\examples{
# extract the invariants from the data
log_ret <- diff(log(EuStockMarkets))

# compute the covariance matrix
sigma <- stats::cov(log_ret)

# torsion
torsion_cov <- torsion(Sigma = sigma, model = 'minimum-torsion', method ='exact')

# 1/N reference
b <- rep(1 / ncol(log_ret), ncol(log_ret))

# ENB
effective_bets(b = b, sigma = sigma, t = torsion_cov)
}
\seealso{
A. Meucci, A. Santangelo, R. Deguest - "Measuring Portfolio Diversification Based on Optimized Uncorrelated Factors" (2013)
}
