% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_col_named}
\alias{is_col_named}
\title{Test if an object has is named}
\usage{
is_col_named(x)
}
\arguments{
\item{x}{The object to be tested.}
}
\value{
A flag (`TRUE` or `FALSE`)
}
\description{
Test if an object has is named
}
\keyword{internal}
