% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_quadratic}
\alias{is_quadratic}
\title{Test if an object is quadratic}
\usage{
is_quadratic(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
A flag (`TRUE` or `FALSE`)
}
\description{
Test if an object is quadratic
}
\keyword{internal}
