% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sqrtm}
\alias{sqrtm}
\title{MATLAB `sqrtm` functionality}
\usage{
sqrtm(x)
}
\arguments{
\item{x}{A square matrix.}
}
\value{
A squared \code{matrix}.
}
\description{
This function returns the square root of a quadratic matrix.
}
\keyword{internal}
