% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sqrtm}
\alias{sqrtm}
\title{Returns the square root of a symmetric positive definite matrix}
\usage{
sqrtm(A)
}
\arguments{
\item{A}{A square matrix}
}
\value{
A quadratic matrix.
}
\description{
This function was copied and pasted from the \code{rrcov} package to avoid dependencies.
}
\keyword{internal}
